+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix argument check against permitted values to allow nil value when optional
 * Fix parsing values provided with abbreviated long option names
 * Fix any object type to symbol conversion failure to return an undefined value
+* Fix parameter validation to stop transforming hash value into an array
 
 ## [v0.2.0] - 2021-08-04